OptionPlay-Web/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
slowapi
pytest
httpx
ruff
nest_asyncio

